Miscellaneous changes
- fmt + clippy
- added StringTag and BoolTag to `ir\values.rs`
- TODO: turn tag comments into correct tags
