+++ src/ir/codeblock.rs
@@ -1,0 +1,9 @@
+use super::{actions::*, events::*};
+
+#[allow(dead_code)]
+pub enum IRCodeBlock {
+    PlayerAction(IRPlayerAction),
+    PlayerEvent(IRPlayerEvent),
+    EntityAction(IREntityAction),
+    EntityEvent(IREntityEvent),
+}
