+++ src/tokengrouper/grouper.rs
@@ -1,6 +1,9 @@
 use itertools::Itertools;
 
-use crate::{lexer::tokens::{Token, TokenType}, ir::IRCodeBlock};
+use crate::{
+    ir::codeblock::IRCodeBlock,
+    lexer::tokens::{Token, TokenType},
+};
 
 #[allow(unused_variables, unused_assignments)]
 /* pub fn transform_to_ast(input: Vec<Token>) -> Vec<Token> {
@@ -79,14 +82,14 @@ use crate::{lexer::tokens::{Token, TokenType}, ir::IRCodeBlock};
     }
     output
 } */
-#[allow(unused_variables, unused_assignments)]
+#[allow(unused_variables, unused_assignments, unused_mut)]
 pub fn transform_to_ast(input: Vec<Token>) -> Vec<Token> {
     let mut output: Vec<Token> = vec![];
     let mut input_iter = input.into_iter();
     // ignore old ast function, this is a new one
     // valid code: parse_event(&input_iter);
     // step 1, parse each code block
-    
+
     // step 2, if you get a successful block, don't continue parsing that and insert the IR variant
     // step 3, if it's an action or anything, check the arguments to make sure it's valid
     // step 4, do custom processing for others
@@ -94,14 +97,13 @@ pub fn transform_to_ast(input: Vec<Token>) -> Vec<Token> {
     output
 }
 
+#[allow(dead_code, unused_variables)]
 pub fn parse_event(iter: &mut dyn Iterator<Item = Token>) -> Option<IRCodeBlock> {
     let mut iter = iter.multipeek();
 
     if let Some(token_1) = iter.peek() {
         if let TokenType::Identifier(ident) = &token_1.token {
-            if let Some(token_2) = iter.peek() {
-
-            }
+            if let Some(token_2) = iter.peek() {}
         }
     }
